Test client-side upload file validation
