+++ public/mock_data/validation/metadata_example_bad_TYPE.txt
@@ -1,0 +1,17 @@
+NAME	Cluster	Sub-Cluster	Average Intensity
+notTYPE	group	group	numeric
+CELL_0001	CLST_A	CLST_A_1	7.742
+CELL_0002	CLST_A	CLST_A_1	-13.745
+CELL_0003	CLST_A	CLST_A_1	-0.375
+CELL_0004	CLST_A	CLST_A_2	9.188
+CELL_0005	CLST_A	CLST_A_2	9.652
+CELL_0006	CLST_B	CLST_B_1	-4.125
+CELL_0007	CLST_B	CLST_B_1	-2.275
+CELL_0008	CLST_B	CLST_B_2	-12.606
+CELL_0009	CLST_B	CLST_B_2	3.675
+CELL_00010	CLST_B	CLST_B_2	-2.116
+CELL_00011	CLST_C	CLST_C_1	0.638
+CELL_00012	CLST_C	CLST_C_1	8.888
+CELL_00013	CLST_C	CLST_C_1	-2.27
+CELL_00014	CLST_C	CLST_C_2	-2.606
+CELL_00015	CLST_C	CLST_C_2	-9.089
